Initial refactor of 'calibrate'
- Split functions into MVC
- Moved printing functions into TerminalPrinter
